export sidecar and roster helpers from cluster_helper
